feat: Move ANNOTATORS to Dialect for dialect-aware annotation
